feat(ai-proxy): allow model & provider to belong to a single client (#6619)

+++ internal/apps/ai-proxy/handlers/handler_model/handler.go
@@ -16,9 +16,12 @@ package handler_model
 
 import (
 	"context"
+	"fmt"
 
+	clientmodelrelationpb "github.com/erda-project/erda-proto-go/apps/aiproxy/client_model_relation/pb"
 	"github.com/erda-project/erda-proto-go/apps/aiproxy/model/pb"
 	commonpb "github.com/erda-project/erda-proto-go/common/pb"
+	"github.com/erda-project/erda/internal/apps/ai-proxy/common/auth"
 	"github.com/erda-project/erda/internal/apps/ai-proxy/providers/dao"
 )
 
@@ -31,21 +34,66 @@ func (h *ModelHandler) Create(ctx context.Context, req *pb.ModelCreateRequest) (
 }
 
 func (h *ModelHandler) Get(ctx context.Context, req *pb.ModelGetRequest) (*pb.Model, error) {
-	return h.DAO.ModelClient().Get(ctx, req)
+	resp, err := h.DAO.ModelClient().Get(ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	// data sensitive
+	desensitizeModel(ctx, resp)
+	return resp, nil
 }
 
 func (h *ModelHandler) Update(ctx context.Context, req *pb.ModelUpdateRequest) (*pb.Model, error) {
-	return h.DAO.ModelClient().Update(ctx, req)
+	resp, err := h.DAO.ModelClient().Update(ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	desensitizeModel(ctx, resp)
+	return resp, nil
 }
 
 func (h *ModelHandler) Delete(ctx context.Context, req *pb.ModelDeleteRequest) (*commonpb.VoidResponse, error) {
+	// check client-model-relation first
+	relationPagingResp, err := h.DAO.ClientModelRelationClient().Paging(ctx, &clientmodelrelationpb.PagingRequest{
+		PageNum:  1,
+		PageSize: 1,
+		ModelIds: []string{req.Id},
+	})
+	if err != nil {
+		return nil, err
+	}
+	if relationPagingResp.Total > 0 {
+		return nil, fmt.Errorf("model is assigned to clients, can not delete")
+	}
 	return h.DAO.ModelClient().Delete(ctx, req)
 }
 
 func (h *ModelHandler) Paging(ctx context.Context, req *pb.ModelPagingRequest) (*pb.ModelPagingResponse, error) {
-	return h.DAO.ModelClient().Paging(ctx, req)
+	resp, err := h.DAO.ModelClient().Paging(ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	// data sensitive
+	for _, item := range resp.List {
+		desensitizeModel(ctx, item)
+	}
+	return resp, nil
 }
 
 func (h *ModelHandler) UpdateModelAbilitiesInfo(ctx context.Context, req *pb.ModelAbilitiesInfoUpdateRequest) (*commonpb.VoidResponse, error) {
 	return h.DAO.ModelClient().UpdateModelAbilitiesInfo(ctx, req)
+}
+
+func desensitizeModel(ctx context.Context, item *pb.Model) {
+	// pass for: admin
+	if auth.IsAdmin(ctx) {
+		return
+	}
+	// if a model is client-belonged, pass for the client
+	if item.ClientId != "" && auth.IsClient(ctx) {
+		return
+	}
+	// hide sensitive data for non-admin
+	item.ApiKey = ""
+	item.Metadata.Secret = nil
 }
